+++ Brainz-App/Brainz-Saves.h
@@ -3,5 +3,5 @@
 
 void fMainSaves();
 void fGetBrains(BRAIN_LIST* brain_list);
-void fSplitBrain(int* id, char* name, char* description, int* is_available, int* note, char* str);
+void fSplitBrain(int* id, char* name, char* description, int* is_available, float* note, char* str);
 void fDisplayBrainList(BRAIN_LIST* brain_list);
